+++ src/lib/waku_message/index.ts
@@ -205,11 +205,7 @@ export class WakuMessage {
       return '';
     }
 
-    return Array.from(this.proto.payload)
-      .map((char) => {
-        return String.fromCharCode(char);
-      })
-      .join('');
+    return Buffer.from(this.proto.payload).toString('utf-8');
   }
 
   get payload(): Uint8Array | undefined {