+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.10.0] - 2021-08-06
+
 ### Added
 - Relay and ReactJS guides and examples
   ([#56](https://github.com/status-im/js-waku/issues/56)).
@@ -163,7 +165,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [ReactJS Chat App example](./examples/web-chat).
 - [Typedoc Documentation](https://status-im.github.io/js-waku/docs).
 
-[Unreleased]: https://github.com/status-im/js-waku/compare/v0.8.1...HEAD
+[Unreleased]: https://github.com/status-im/js-waku/compare/v0.10.0...HEAD
+[0.10.0]: https://github.com/status-im/js-waku/compare/v0.9.0...v0.10.0
+[0.9.0]: https://github.com/status-im/js-waku/compare/v0.8.1...v0.9.0
+[0.8.1]: https://github.com/status-im/js-waku/compare/v0.8.0...v0.8.1
 [0.8.1]: https://github.com/status-im/js-waku/compare/v0.8.0...v0.8.1
 [0.8.0]: https://github.com/status-im/js-waku/compare/v0.7.0...v0.8.0
 [0.7.0]: https://github.com/status-im/js-waku/compare/v0.6.0...v0.7.0